Separated StocksDashboard and Formatter into different files
