+++ stocksdashboard/__init__.py
@@ -1,6 +1,12 @@
 from .stocksdashboard import StocksDashboard
-from .stocksdashboard import Formatter
 from .stocksdashboard import convert_to_datetime
 from .stocksdashboard import get_colors
+from .formatter import Formatter
 
-__all__ = ['StocksDashboard', 'Formatter', 'convert_to_datetime', 'get_colors']
+__all__ = ['StocksDashboard', 'convert_to_datetime', 'get_colors', 'Formatter']
+
+
+__author__ = ["Mabel Villalba Jimenez (mabelvj@gmail.com)",
+              "Emilio Molina Martinez (emilio.mol.mar@gmail.com)"]
+__license__ = "GPLv3"
+__version__ = '0.1dev0'
